\name{GetNumberOfPedigrees}
\alias{GetNumberOfPedigrees}
\title{
Get Number of Pedigrees. 
}
\description{Get the number of pedigrees stored in Familias. 
}
\usage{
GetNumberOfPedigrees()
}
\value{The number of pedigrees stored in Familias. 
}
\author{
Petter Mostad <mostad@chalmers.se>
}
\seealso{
\code{\link{GetPedigree}}
}
\examples{
GetNumberOfPedigrees()
}
