\name{NewFamilias}
\alias{NewFamilias}
\title{Removes all stored information from Familias 
}
\description{
Removes all stored information, including information about persons, 
allele systems, and pedigrees. 
}
\usage{
NewFamilias()
}
\value{
None
}
\author{
Petter Mostad <mostad@chalmers.se> 
}
\examples{
NewFamilias()
}
